webapi/tct-pagevisibility-w3c-tests/pagevisibility/bdd/pagevisibility_DocumentVisibility_hidden_minimize-manual.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        HaoYunfen <yunfenx.hao@intel.com>
        Xu,Yuhan <yuhanx.xu@intel.com>

-->

<html>
    <head>
        <meta content="text/html; charset=utf-8" http-equiv="Content-Type" />
        <title>Check if hidden attribute is false while the webpage from minimize change to visible</title>
        <SCRIPT src="../../resources/testharness.js"></SCRIPT>
        <script src="../../resources/testharnessreport.js"></script>
        <link rel="stylesheet" href="../../resources/testharness.css" />
    </head>
    <body>
        <div id="log"></div>
        <script>
            setTimeout(hiddenTest,4000);
            var t=async_test("Check if hidden attribute is false while the webpage from minimize change to visible");
            function hiddenTest(){
                var hidden = document.hidden;
                t.step(function() {
                    assert_true(!hidden,"expects hidden attribute is false");
                });
                t.done();
            }
        </script>
    </body>
</html>
